Store the offset of the task name in the tcb for openocd debugging
Previously the task name was 52 bytes from the start of the structure.
Now with the SMP branch, there are various options that will cause
this to shift based on whatever is enabled.

We could force the task name to be the second entry in the structure
to ensure it is always as the same spot but this would break all
existing openocd implementations.

We could restructure the tcb structure to move the new additions
below the task name, but that's probably just delaying the problem
until it happens again.

Instead, simply store the offset so that openocd can look it up at
runtime.